requirements.txt
yfinance
pandas
matplotlib
numpy
requests
python-dotenv
scipy